Move classes out of Renderer to separate headers
